Bug que impedia a criação da notificação resolvido.
