+++ app/src/main/java/com/example/appbebergua/activity_alarmes.java
@@ -1,9 +1,14 @@
 package com.example.appbebergua;
 
-import static com.example.appbebergua.Timer.*;
+import static com.example.appbebergua.Notificacao.*;
 import androidx.appcompat.app.AppCompatActivity;
+
+import android.app.AlarmManager;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.os.Bundle;
 import android.provider.AlarmClock;
 import android.view.View;
@@ -11,31 +16,28 @@ import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.RadioButton;
-import android.widget.RadioGroup;
 import android.widget.Toast;
 
 public class activity_alarmes extends AppCompatActivity {
 
     private EditText edtHoraOuMinuto, edtMinutoOuSegundo;
-    private Button btnSetAlarm;
     private RadioButton rdbAlarme, rdbTimer;
-    private RadioGroup rdgNotificacao;
     private String escolhaUsuario;
-    private boolean alarmeProgramado;
-    private boolean timerRodando;
+    //private boolean alarmeProgramado;
+    private Notificacao notificacao;
+    private Button btnSetAlarm;
 
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_alarmes);
 
-        setContext(this);
+        createNotificationChannel();
         edtHoraOuMinuto = findViewById(R.id.edtHoraOuMinuto);
         edtMinutoOuSegundo = findViewById(R.id.edtMinutoOuSegundo);
         btnSetAlarm = findViewById(R.id.btnSetAlarm);
         rdbAlarme = findViewById(R.id.rdbAlarme);
         rdbTimer = findViewById(R.id.rdbTimer);
-        rdgNotificacao = findViewById(R.id.rdgNotificacao);
 
         rdbAlarme.setOnClickListener(view -> {
             edtHoraOuMinuto.setHint("Digite a hora.");
@@ -45,46 +47,52 @@ public class activity_alarmes extends AppCompatActivity {
         rdbTimer.setOnClickListener(view -> {
             edtHoraOuMinuto.setHint("Digite os minutos.");
             edtMinutoOuSegundo.setHint("Digite os segundos.");
-
         });
 
-
-        btnSetAlarm.setOnClickListener(view -> {
+        btnSetAlarm.setOnClickListener(view ->{
+            AlarmManager alarmManager = (AlarmManager) getSystemService(ALARM_SERVICE);
+            notificacao = new Notificacao(this, alarmManager);
+            notificacao.setNotificationAlarm();
             closeKeyboard();
             if (rdbAlarme.isChecked()) {
-
-                int hour = Integer.parseInt(edtHoraOuMinuto.getText().toString());
-                int minute = Integer.parseInt(edtMinutoOuSegundo.getText().toString());
-
-                Intent intent = new Intent(AlarmClock.ACTION_SET_ALARM);
-                intent.putExtra(AlarmClock.EXTRA_HOUR, hour);
-                intent.putExtra(AlarmClock.EXTRA_MINUTES, minute);
-                intent.putExtra(AlarmClock.EXTRA_MESSAGE, "Tomar água");
-
-                if (correctInputAlarme(hour, minute)) {
-                    escolhaUsuario = "alarme";
-                    startActivity(intent);
-                }
+                setAlarme();
             } else if (rdbTimer.isChecked()) {
-                String inputMinutos = edtHoraOuMinuto.getText().toString();
-                String inputSegundos = edtMinutoOuSegundo.getText().toString();
-                if(correctInputTimer(inputMinutos, inputSegundos)) {
-                    escolhaUsuario = "timer";
-                    if (isTimerRunning()) {
-                        Toast.makeText(this, "Timer já definido!", Toast.LENGTH_LONG).show();
-                    } else {
-                        startTimer();
-                    }
-                }
+                setTimer();
             }
         });
     }
 
+    public void setAlarme(){
+        int hour = Integer.parseInt(edtHoraOuMinuto.getText().toString());
+        int minute = Integer.parseInt(edtMinutoOuSegundo.getText().toString());
+
+        Intent intent = new Intent(AlarmClock.ACTION_SET_ALARM);
+        intent.putExtra(AlarmClock.EXTRA_HOUR, hour);
+        intent.putExtra(AlarmClock.EXTRA_MINUTES, minute);
+        intent.putExtra(AlarmClock.EXTRA_MESSAGE, "Tomar água");
+
+        if (correctInputAlarme(hour, minute)) {
+            escolhaUsuario = "alarme";
+            startActivity(intent);
+        }
+    }
+
+    public void setTimer(){
+        String inputMinutos = edtHoraOuMinuto.getText().toString();
+        String inputSegundos = edtMinutoOuSegundo.getText().toString();
+        if(correctInputTimer(inputMinutos, inputSegundos)) {
+            escolhaUsuario = "timer";
+            if (isTimerRunning()) {
+                Toast.makeText(this, "Timer já definido!", Toast.LENGTH_LONG).show();
+            } else {
+                setEndTime();
+                notificacao.setNotificationAlarm();
+            }
+        }
+    }
+
     public boolean correctInputAlarme(int hour, int minute){
-        if(hour <= 24 && hour > 0 &&  minute <= 60)
-            return true;
-        else
-            return false;
+        return hour <= 24 && hour > 0 && minute <= 60;
     }
 
     //Métodos para o timer
@@ -113,10 +121,28 @@ public class activity_alarmes extends AppCompatActivity {
             return false;
         }
 
-        setTime(millisInputMinutos, millisInputSegundos);
+        setNotificationTime(millisInputMinutos, millisInputSegundos);
         edtHoraOuMinuto.setText(null);
         edtMinutoOuSegundo.setText(null);
         return true;
+    }
+
+    //Métodos para a notificação
+    public void createNotificationChannel(){
+        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            CharSequence name = "drinkWaterAppReminderChannel";
+            String description = "Channel to notify the time to drink water";
+            int importance = NotificationManager.IMPORTANCE_DEFAULT;
+            NotificationChannel channel = new NotificationChannel("drinkWaterApp", name, importance);
+            channel.setDescription(description);
+
+            NotificationManager notificationManager = getSystemService(NotificationManager.class);
+            notificationManager.createNotificationChannel(channel);
+        }
+    }
+
+    public void setNotificationTime(long milliSecondsMinutes, long milliSeconds){
+        setStartTimeInMillis(milliSecondsMinutes + milliSeconds);
     }
 
 
@@ -138,20 +164,7 @@ public class activity_alarmes extends AppCompatActivity {
         SharedPreferences prefs = getSharedPreferences("prefs", MODE_PRIVATE);
 
         setStartTimeInMillis(prefs.getLong("startTimeInMillis", 10000));
-        setTimeLeftInMillis(prefs.getLong("millisLeft", getStartTimeInMillis()));
         setTimerRunning(prefs.getBoolean("timerRunning", false));
-
-        if (isTimerRunning()) {
-            setEndTime(prefs.getLong("endTime", 0));
-            setTimeLeftInMillis(getEndTime() - System.currentTimeMillis());
-
-            if (getTimeLeftInMillis() < 0) {
-                setTimeLeftInMillis(0);
-                setTimerRunning(false);
-            } else {
-                startTimer();
-            }
-        }
     }
 
     //Método que "salva" os valores quando o app é fechado
@@ -163,13 +176,10 @@ public class activity_alarmes extends AppCompatActivity {
         SharedPreferences.Editor editor = prefs.edit();
 
         editor.putLong("startTimeInMillis", getStartTimeInMillis());
-        editor.putLong("millisLeft", getTimeLeftInMillis());
+        //editor.putLong("millisLeft", getTimeLeftInMillis());
         editor.putLong("endTime", getEndTime());
         editor.putBoolean("timerRunning", isTimerRunning());
         editor.apply();
 
-        if (getCountDownTimer() != null) {
-            cancelCountDownTimer();
-        }
     }
 }
